+++ src/main/java/dev/jojofr/joseta/utils/BotCache.java
@@ -3,7 +3,7 @@ package dev.jojofr.joseta.utils;
 import dev.jojofr.joseta.JosetaBot;
 import dev.jojofr.joseta.database.Database;
 import dev.jojofr.joseta.database.daos.ConfigurationDao;
-import dev.jojofr.joseta.database.daos.MessageDao;
+import dev.jojofr.joseta.database.daos.messages.MarkovBlacklistDao;
 import dev.jojofr.joseta.database.entities.ConfigurationEntity;
 import dev.jojofr.joseta.entities.GuildConfiguration;
 import net.dv8tion.jda.api.entities.emoji.Emoji;
@@ -34,7 +34,7 @@ public class BotCache {
             }
             
             GuildConfiguration guildConfig = new GuildConfiguration(config);
-            guildConfig.markovBlacklistIds = Database.withExtension(MessageDao.MarkovBlacklistDao.class, dao -> dao.getAllIds(id));
+            guildConfig.markovBlacklistIds = Database.withExtension(MarkovBlacklistDao.class, dao -> dao.getAllIds(id));
             
             return guildConfig;
         });
